Rearrange files and add parallel optimization solver class which wraps around yaw_optimizer_wind_rose_serial.py. This really cleans up the yaw optimization codes and minimizes overlap between different files while keeping all the benefits of parallelization.
