+++ floris/tools/optimization/__init__.py
@@ -1,1 +1,1 @@
-from . import other, scipy, pyoptsparse, farm_cost_function, yaw
+from . import other, scipy, pyoptsparse, yaw_optimization
