--- floris/tools/optimization/general_library/__init__.py
@@ -1,1 +1,0 @@
-from . import farm_cost_function, optimization, yaw_scipy, yaw_sr, derive_downstream_turbines
